add prompt

+++ Resizer/BaseResizer.cs
@@ -1,26 +1,47 @@
+using NeoSmart.PrettySize;
 public abstract class BaseResizer
 {
     protected CMPDF.PDF pdf = new CMPDF.PDF();
     protected byte[] _oriFileBytes;
     protected string _oriFileName;
-    protected FileInfo? _outputFile;      
+    protected FileInfo? _outputFile;
     protected BaseResizer(FileInfo oriFile, FileInfo? outputFile)
     {
         _oriFileBytes = File.ReadAllBytes(oriFile.FullName);
         _oriFileName = oriFile.FullName;
         _outputFile = outputFile;
     }
-    abstract public byte[] resize(int maxWidth=800);
-    public void writeFile(byte[] newFile)
+    abstract public byte[] resize(int maxWidth = 800);
+    public void writeFile(byte[] newFile, bool Prompt)
     {
         if (newFile.Length < _oriFileBytes.Length)
         {
-            File.WriteAllBytes(_outputFile != null ? _outputFile.FullName : _oriFileName, newFile);
-            Console.WriteLine(string.Format("{0} {1} completed {2}",
-            _outputFile != null ? "Write" : "Overwrite",
-            _outputFile != null ? _outputFile.Name : Path.GetFileName(_oriFileName),
-            efficiency(_oriFileBytes.Length, newFile.Length))
-            );
+            var newSize = PrettySize.Megabytes(newFile.Length);
+            var oriSize = PrettySize.Megabytes(_oriFileBytes.Length);
+            string? toWrite;
+            if (Prompt)
+            {
+                Console.Write(string.Format("Do you want to {0} to the file({1})? [Y/N]",
+                _outputFile != null ? "Write" : "Overwrite", newSize));
+                toWrite = Console.ReadLine();
+            }
+            else
+                toWrite = "Y";
+
+            if (!string.IsNullOrEmpty(toWrite) && toWrite.ToUpper() == "Y")
+            {
+                File.WriteAllBytes(_outputFile != null ? _outputFile.FullName : _oriFileName, newFile);
+                Console.WriteLine(string.Format("{0} {1} completed {2}",
+                _outputFile != null ? "Write" : "Overwrite",
+                _outputFile != null ? _outputFile.Name : Path.GetFileName(_oriFileName),
+                efficiency(_oriFileBytes.Length, newFile.Length))
+                );
+            }
+            else
+            {
+                Console.WriteLine("Cancel writing to the file");
+            }
+
         }
         else
         {
@@ -30,7 +51,9 @@ public abstract class BaseResizer
 
     private string efficiency(int ori, int com)
     {
-        return $"[{(int)(((double)(ori - com) / ori) * 100)}%]";
+        var newSize = PrettySize.Megabytes(com);
+        var oriSize = PrettySize.Megabytes(ori);
+        return $"[{newSize}/{oriSize}] [{(int)(((double)(ori - com) / ori) * 100)}%]";
     }
 
 }
